+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added support for different prediction types
 - Added support for hashes and Rover data frames to `predict` method
 - Added support for hashes to `Dataset`
+- Added `importance_type` option to `dump_model`, `model_to_string`, and `save_model` methods
 - Changed `Dataset` to use column names for feature names with Rover and Daru
 - Changed `predict` method to match feature names with Daru
 - Dropped support for Ruby < 3.1